djangoapp/requirements.txt
Django==5.2.6
httpx==0.27.2
python-dotenv==1.0.1